requirements.txt
python-docx==1.1.2
reportlab==4.2.5
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0